backend/requirements.txt
fastapi[standard]==0.115.6
google-generativeai==0.8.3
python-dotenv==1.0.1